requirements.txt
requests
python-dotenv
anthropic
pymupdf
python-docx
openpyxl
openai-whisper
mcp
openai
